requirements.txt
fastapi
uvicorn
pymongo
python-jose
bcrypt
python-multipart
python-dotenv